raptor/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "raptor-adaptation"
version = "1.0.0"
authors = []
description = ""
readme = ""
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
    "faiss-cpu",
    "numpy==1.26.3",
    "openai==1.3.3",
    "scikit-learn",
    "sentence-transformers~=3.0.1",
    "tenacity==8.2.3",
    "tiktoken==0.5.1",
    "torch",
    "transformers==4.38.1",
    "umap-learn==0.5.5",
    "urllib3==1.26.6",
    "llama-cpp-python~=0.2.79",
    "llama-index~=0.10.38",
]

[tool.setuptools]
py-modules = []
